+++ envlib/__init__.py
@@ -4,7 +4,7 @@ from envlib import vocabularies
 from envlib.catalogue import Catalogue, DatasetRef
 from envlib.metadata import Metadata, ValidationError, canonical_station_point, compute_station_id
 
-__version__ = '0.1.4'
+__version__ = '0.1.5'
 
 __all__ = [
     'Catalogue',